Fix FFI seeds type

+++ ffi/src/lib.rs
@@ -1,13 +1,13 @@
 use core::slice;
 
 #[no_mangle]
-pub unsafe extern "C" fn gxhash32(buf: *const (), len: usize, seed: i32) -> u32 {
+pub unsafe extern "C" fn gxhash32(buf: *const (), len: usize, seed: i64) -> u32 {
     let data: &[u8] = slice::from_raw_parts(buf as *const u8, len);
     gxhash::gxhash32(data, seed)
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn gxhash64(buf: *const (), len: usize, seed: i32) -> u64 {
+pub unsafe extern "C" fn gxhash64(buf: *const (), len: usize, seed: i64) -> u64 {
     let data: &[u8] = slice::from_raw_parts(buf as *const u8, len);
     gxhash::gxhash64(data, seed)
 }
